Restore the detekt gate to green
detektAll failed with 11 unbaselined issues, most of them genuine
complexity regressions in the recovery code. Resolve them
structurally rather than by baseline growth:

- Split submitSingleDelegationBundle into the cached-hash probe,
  submission resolution, and confirmed-transaction storage steps.
- Extract the spent-nullifier recovered-delegation resolution from
  reconcileDelegationTransactionResult.
- Move the commitment-tree scanner and its validation helpers into
  CommitmentTreeScan.kt.
- Collapse castVoteLeafPositions's per-part throws into one malformed
  leaf_index error.
- Fold existingRoundRecoveryAction's catch into the classifying when.

The baseline loses the stale entry for the deleted isRoundRecoverable
and gains only LargeClass for the progress test suite.
